Time: 1605 ms (7.56%), Space: 263.4 MB (27.01%) - LeetHub

+++ 0446-arithmetic-slices-ii-subsequence/0446-arithmetic-slices-ii-subsequence.cpp
@@ -8,7 +8,7 @@ class Solution {
 
       for (int i = 1; i < n; i ++) {
         for (int j = i - 1; j >= 0; j--) {
-          long diff = (long)nums[i] - (long)nums[j];
+          long long diff = (long long)nums[i] - (long long)nums[j];
           count += dp[j][diff];
